langraph-agents/requirements.txt
# LangGraph and LangChain Core Dependencies
langgraph>=0.2.30
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# Web Framework and API Dependencies
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2
requests==2.32.3
aiohttp==3.10.10

# Document Processing and Office Integration
python-docx==1.1.2
openpyxl==3.1.5
PyPDF2==3.0.1
python-multipart==0.0.12

# Data Processing and Analysis
pandas==2.2.3
numpy>=1.26.4,<2.0.0
matplotlib==3.9.2
plotly==5.24.1

# AI and Machine Learning
openai==1.51.2
anthropic==0.34.2
tiktoken==0.8.0
sentence-transformers==3.2.1

# Database and Storage
sqlalchemy>=1.4,<2.1
# sqlite3 is built-in with Python
redis>=5.0.0
chromadb>=0.5.0

# Configuration and Environment Management
python-dotenv==1.0.1
pyyaml==6.0.2
toml==0.10.2

# Logging and Monitoring
structlog==24.4.0
rich==13.9.2

# Testing and Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.10.0
isort==5.13.2
mypy==1.13.0

# Utilities
typing-extensions==4.12.2
python-dateutil==2.9.0.post0
jinja2==3.1.4

# Financial Data APIs (for finance-office integration)
yfinance==0.2.38
alpha-vantage==2.3.1
quandl==3.7.0

# Security and Authentication
cryptography==43.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# XML and JSON Processing
lxml==5.3.0
xmltodict==0.14.2
jsonschema==4.23.0

# Task Queue and Background Processing
celery==5.4.0
kombu==5.4.2

# Network and HTTP Tools
urllib3==2.2.3
certifi==2024.8.30

# Development and Debugging
ipython==8.29.0
jupyter==1.1.1